Use pre-compiled 'grep' binary from MSYS2 on Windows
This drops the self-compiled 'grep' version which caused issues
with recursive searches on Windows as well as with non-ASCII
characters in paths.
The pre-compiled 'grep' binary from MSYS2 (matching the target ABI)
seems to solve probably all of the known problems on Windows.
The only drawback is that we need to ship a few additional MSYS2
libraries on which the 'grep' binary depends.

Closes #1229 and #1260.