scintilla: Prevent running signal handlers on a destroyed a11y object
Avoid crash when detaching the widget from the accessible object
without destroying that widget.

In such situations, the widget is still valid but we will have
destroyed the orphaned accessible object.  Thus, we must make sure we
disconnected the signal handlers the late accessible had set up on the
widget, as they won't be implicitly disconnected by widget
finalization in this case.

Fixes #1385.