Build system preparations for using the Criterion unit testing framework
Criteron is a "dead-simple, yet extensible, C and C++ unit testing framework.".
It's really simple to use, but still a powerful framework for writing unit
tests in C.

This commit adds an m4 function that checks for criterion unit test framwork
and a wrapper script for running the unit tests under the Automake TAP
driver.

See https://github.com/Snaipe/Criterion for an overview, and specifically
https://github.com/Snaipe/Criterion/blob/bleeding/dev/autotools/README.md
for autotools integration.