snippets: Remove cursor position at the end of constructs.
This is not consistently done for all languages, and hard to get right
e.g. for python. It's probably not terribly useful either.

This also interacts bad if snippet-next-cursor is mapped to the Tab key,
in which case the cursor position will prevent to use Tab for indentation
inside inserted blocks:

if (<cursor1>)
{
    <cursor2> /* want to use Tab for indentation here */
}
<cursor3>

The easiest way to go to remove cursor3 from the snippet.