api: Increment API version.
The snippets improvements make editor_insert_snippet() more attractive to
plugins (I certainly want to use it now), so make allow plugins to detect this.

Also, the allocation of the scintilla indicator for snippets may be important
to know about.

ABI need not be incremented, the improvements aren't expected to break
anything such that recompiling would be sufficient.