snippets: fix start/end detection, when searching for the next cursor
Start of file wasn't properly detected when it started with a cursor marker,
because the logic attempted to look before that marker.

End of file wasn't properly detected was faulty and made the cursor jump to the
end of the doc. Now, EOF is detected by a zero-length marker which can't happen
in the middle of a document.