plugin-manager: sync buttons on "button-release-event"
On fast-clicks/double-clicks on a plugin row the selected row can change
and the buttons (e.g. the "Help" button) can get out of sync. That means
that e.g. a plugin is selected which has not got an help function but
the "Help" button is activated. Clicking the button is such a situation
leads to a crash. Syncing buttons on "button-release-event" solves the
problem. Fixes #1781.