Fix plugin manager UI synchronization on double click
When toggling a plugin, we temporarily set the tree store's row entry
for the plugin pointer to NULL as we destroy and reload the selected
plugin, and its pointer would be invalid in the meantime.  This results
in the filter we use to display search results to temporarily hide the
row, changing the actual number of rows and thus, depending on timing,
this will or will not change the selected row (it will when double
clicking, not when single-clicking), in a seemingly more or less random
fashion as we use a sorted model.

Finally, as we manually update the buttons visibility for the toggled
plugin (as we otherwise do only for changing selection, which should
not happen in this case -- well, most of the time as you can see), this
can lead to the buttons to be updated for a now unselected row, getting
those out of sync.

The fix here is not to actually hide rows with a NULL plugin, because
it can only happen in 2 cases, where we actually want to see it:

1. while toggling a plugin, as explained above, in which case it had to
   match the search already.
2. when there is no plugins and we want to display a "No plugins
   available" message, and the search should not affect this.

This incidentally also fix the "No plugins available" so it's actually
visible, instead of always hidden.

Fixes #1781.