find-in-files: choose reg-exp mode (extended/basic/perl)
This adds a combo-box to the find-in-files dialog which lets the user choose
which kind of regular expressions shall be used: extended, basic or perl.

The pre-selected value is 'extended' so users which do not want to change
the reg-exp mode used, do not need to change anything. If a user only
toggles the checkbox to enable/disable regular expressions then anything
works as it used to.

Also see #443.