Reuse the functionality of createTagsWithFallback()
By calling createTagsWithFallback() we can reuse some of its functionality
and simplify geanyCreateTagsWithCallback(). Also rename
geanyCreateTagsWithCallback() to geanyCreateTags().

In the future we still have to keep geanyCreateTags() in sync with
parseFileWithMio() but because this function does things we don't want
in Geany it's more practical to do it this way.