Check also regex parser kind mappings
With the past versions of ctags we didn't have access to regex parser
kinds so we couldn't check if the Geany mappings are correct. This
isn't the case any more so we can do the same with regex parsers like
with other parsers and as a bonus we can eliminate one more diff from
ctags.