Make Geany informed about failed reparses "the right way"
At the moment we pass a callback to ctags which informs us about
re-parses and based on which we clear our tag list when a re-parse
happens. The current solution isn't very clean because it introduces
new callback function and many diffs into parse.c and in addition it
doesn't do the correct thing when re-parses happen within subparsers.

Instead, we can introduce a new callback function into writers, use this
function and also get informed about how many tags are valid when a
reparse happens and keep only that number in our array list.