Move ctags-api.c/h into tagmanager and rename it to tm_ctags.c/h
Since we now don't define any types in ctags-api.h that would be used
inside ctags, we can move it to tagmanager so there's not this extra
file in the ctags directory. The real "ctags API" are the internal
functions ctags uses and which we use for our purposes and tm_ctags.c/h
just encapsulate ctags so we don't include ctags headers directly
in the rest of Geany.