Move lcpp.c/h inside parsers dir
This corresponds to cpreprocessor.c/h of upstream ctags which lives
in the parsers directory too. Let's keep the "lcpp" name though - if
we update to the latest cxx parser, we'll have to keep both implementations
of the preprocessor because of the big diffs between our c.c and
ctags c.c.