Inject our parser list to ctags
This patch syncs to ctags version

869dd9d750d08c4bfb8a0aaa38ab84ebc33b222c

which allows us to define a custom parser list. This is done using the
EXTERNAL_PARSER_LIST_FILE (done inside ctags/Makefile.am) which has to
point to the file with the parser list where the parsers have to be
defined inside the EXTERNAL_PARSER_LIST define.

This patch adds new tm_parsers.h header where our parsers are defined.
This means that we don't have to care about ctags/parsers_p.h any more
and it can have the same contents as the upstream ctags.

Thanks to using our parser list directly, we can avoid the "invisible"
test parser in ctags so we don't have to shift parser numbers by 1
any more so the CTAGS_LANG() and GEANY_LANG() macros can be dropped.

Finally the patch updates various comments and the HACKING file to
mention tm_parsers.h instead of ctags/parsers.h. In the HACKING file
it also drops the section about using GRegex which isn't the case any
more.