Rename pointerOrder TMtag member to flags
We now use varType instead of pointerOrder and it has been the case for
quite and we can use the pointerOrder field for storing something useful.
When we store something else inside the renamed pointerOrder, we will
break compatibility with some very old tag files but I don't think there
are any such files around (also pointerOrder was probably just reported
by the C parser).

Bump API for the pointerOrder change to flags