Allow configuring shell used for build commands
Currently there is a problem when users use a different shell than sh/bash
and configure their environment in the config file of the given shell.
This config file isn't loaded by Geany because Geany always uses sh
in a non-login non-interactive way.

This patch tries to solve the problem by allowing users to specify the
shell they want to use for launching build commands. In order to make
shells load their config files, the shells are run as login non-interactive
shells because non-login non-interactive shells like sh/bash don't load
configuration from any files. Unfortunately csh/tcsh don't allow us to be
run as login and use the -c option simultaneously so these shells aren't
started as login shells.

Running shells as login is absolutely necessary on macOS even for sh/bash.
While on linux display managers typically source .profile upon login,
this doesn't happen on macOS and the only way to get the environment loaded
is to run shells as login.

There are some places where /bin/sh can still be used. One of them is
the execute command. When configured to use VTE for execution, we already
have the environment set up by the shell in VTE so we don't need to do any
other work. When configured to use external terminal for running, the
terminal is run as non-login interactive so it should load its
configuration too.

I kept the printing command to use /bin/sh because either the code used
for the build command would have to be duplicated there or somehow
factored-out and it doesn't seem to be worth the work for this simple
case where only /usr/bin/lpr is called.

This patch has been tested with the following shells:
sh, bash, dash, ksh, zsh, csh, tcsh, fish, ion