Make sure any shell used in VTE is run as login shell on macOS
Previously we did this for bash only but we should do it for arbitrary
shell. The -l option seems to be supported by all major shells.

Previously we used to add -l to the settings value itself which probably
isn't the cleanest way as the entry should contain path. Let's add
it invisibly to user but make sure we don't add it when settings already
contains -l.