Remove explicit static accessible name from the main window
Users expect accessible name of the window to match the window title,
so remove the custom explicit name of "Geany" which was added at the
very beginnings of Geany, presumably without real a11y reasons.

Without this, GTK forwards the window title to the a11y name for us,
leading to the expected result and matching most other apps.

Fixes #2421.