Read default browser command using GIO
Instead of using only hard-coded and varying defaults for Linux,
MacOS and Windows, try to read the system's default command for "http"
URI schemes and use it if found.
If no usable default is available, fallback to the previous hard-coded
defaults.

Besides being more user-friendly, this should solve issues on Windows
where we always used "ShellExecute" and bypassed the configured browser
command which led to errors on URLs with anchors (see #2405).