Use Scintilla's JSON lexer for JSON instead of the JavaScript/C++ lexer
This commit changes JSON from a custom filetype to one with its own lexer, as requested in
https://github.com/geany/geany/issues/384#issuecomment-583869454. The lexer is taken from Scintilla's upstream
(version 3.10.4).

This has the following effects:
 - JavaScript comments (//, /* */) are no longer highlighted as comments, since JSON doesn't allow them
 - Text outside of quotes, except keywords and numbers, is now considered an error, and highlighted as such (in red)

It should also allow keys and string values to be highlighted in different colours, as in many other editors, but I couldn't get that to work.

Stopping comments from being highlighted is the main motivation for this - highlighting them misleads the user into thinking that JSON allows them.