Do not convert filename when opening files from command line on Windows
It's unclear why the previous was added (by me) 12 years ago.
Maybe the code in main_handle_filename() or after has been
changed/improved so that it works now or the previous test case for
opening files from the command line was wrong.
This happens only when starting a new instance of Geany from command
line. Opening files in a running instance works fine with and without
this change.

Closes #2652.