Add the geany_ prefix to all parser files
The geany_ prefix is meant to indicate that the parser either has a
geany-specific implementation or that there are some things in the parser
which should be committed upstream. Once a parser is fully synced
with the upstream version, we can drop the geany_ prefix. This helps us
know which parsers are synced (and for which we can just copy the
uctags version) and where we still have some work to do.

In this initial commit, add the geany_ prefix everywhere - some parsers
are already fully synced and the geany_ prefix can be dropped when we
update to the up-to-date uctags version.