Convert working directory from UTF-8 to locale on Windows
When passing filenames from the command line on Windows, the filename
is in system locale. If it is a relative path, we make it absolute with
the current working directory which is in UTF-8 encoding.
So we need to convert the current working directory also to system
locale, join the filename and then convert both back to UTF-8.

Closes #2652.