Consume closing brackets, braces, parenthesis, single and double quotes
For a flowing writing and coding I love to not change my hand position
to use the arrow keys!
I realy like auto-closing.

If I activates the auto closing brackets, braces, parenthesis and
quotes feature, the closing quotes are inserted, but to continue after
these, the arrow keys are necessary.

The modification I made: if autoclosing is activated in
options, if a closing character is typed and the following character
matches this, not both characters stay - one deleted - curr pos
afterwards.

Example: the | marks the curr position:

"|"
"hallo|"
[Then the " is hit]
"hallo"|

The " as well as the other affected characters are reachable without
changing hand pos.

I saw this in other IDEs and I want it in my Geany.