Sync ctags parsers with small changes with Geany (#2990)
* Remove the geany_ prefix from parsers with trivial changes against ctags

This will allow the update-ctags.py script to replace these parsers
with parsers from uctags.

* Update parsers to the uctags versions and update Geany where needed

This patch copies over parsers without the geany_ prefix from uctags
tag p5.9.20211031.0.

Next, it updates our mapping of kinds to tag types when there was
a change. The deleted kind in ABC mapping was never generated by the
ABC parser and was removed in the upstream version. The new 'C' and 'd'
kinds of HAXE parser were mapped to tm_tag_undef_t.

Finally, txt2tags uses a different context separator (\"\" instead of \x03)
so tm_parser_context_separator() was updated accordingly and so were
the unit tests. At the same time I removed TM_PARSER_ASCIIDOC from
tm_parser_context_separator() because the parser doesn't use context
separator at all and it's unnecessary there.

* Add julia test to Makefile.am

This test was missing and the unit test wasn't performed.

* Remove REST from tm_parser_context_separator()

The mentioned scope separator ":::" isn't used in the REST parser. The
parser only includes one-level scope (only immediate parent) so
scope separators aren't used for it.

* Undo previous commit removing REST scope separator and improve comment

* Add asciidoc among languages without nested scope

It was removed in the initial commit but even though it doesn't use
nested scope, '.' could appear in the title so use something less probable.

* Use "\x3" as the improbable character for scope separator

It was used previously as a txt2tags scope separator and since it's
unprintable, it's better than ::: that could be written in the title.

On the way, move the real "\"\"" scope separator above in the switch
list so the special "\x3" case is listed last.