Add a mechanism to enable/disable some ctags kinds
This patch is similar to the patch allowing us to enable/disable roles,
it just does this for ctags kinds. This can be useful when there is
a bug in a ctags parser causing incorrect parsing when certain kind is
enabled - which is what happens when parsing cython source files
with the 'z' flag enabled.