Make the New Project dialog a little wider
With the current width, the created project file name gets hidden even
for not-so-long paths. Consider for instance the following path:

/home/my_name/projects/project/project.geany

It consists of 43 characters where "/home/my_name/projects/" is a typical
prefix of someone with not terribly long user name and also
"project/project.geany" is fairly normal project name and directory name
length. Even though the patch says 40, this 43-character string still fits
in on my machine (50 makes the window a little too wide).