Pass our ignore.tags file with ignored symbols to cxx preprocessor
The syntax is slightly different from the previous syntax and is
described here:

https://docs.ctags.io/en/latest/parser-cxx.html

Basic usage should be the same, uctags just doesn't support Geany's
wildcard ignores like G_GNUC_*. On the other hand the new parser is
much more resilient to macros so there shouldn't be so much need
for manual ignores.

The original code is still kept for parsers from c.c that still use the
old preprocessor.