Sync ctags parsers with big changes with Geany (#2991)
* Remove the geany_ prefix from selected parsers and use the corresponding uctags version

* Update tag mappings in tm_parser.c

* Update PHP and zephir context separator to "\\" which is used in uctags now

* Update ruby unit tests, the tests only differ in added "()" to all functions in generated tags.

* Update nsis unit tests, the section name seems to be parsed correctly now.

* Update go unit tests, the new go tags contain extra scope information.

* Update php and zephir unit tests, PHP and zephir now use "\\" as the context separator instead of "::".
For zephir, return values of methods are now parsed.

* Update Objective-C unit tests, the new parser parses also method arguments.

* Update lua unit tests, Scope information seems to be correctly parsed now.

* Add uctags iniconf parser

* Update tag mappings for the iniconf parser

* Add a way to modify scope information if needed and use it to unify php separators

This patch introduces a new function tm_parser_update_scope() which
can be used to modify scope information when the provided value from
ctags doesn't suit our needs.

This patch uses it to replace php scope separator \ with :: so there's
a single scope separator for this language.

* Remove comment regarding ruby scope separator

"." is used as the context separator in ruby now, the comment is
probably just some historic artifact.

* Add comment explaining why we modify PHP scope separators

* Add a mechanism to enable/disable roles for certain kinds

This patch allows us to enable/disable ctags roles for certain languages
and kinds. Roles are currently disabled only for the Go kind 'p'
for the reasons mentioned in the comment message in the patch.

* Add a mechanism to enable/disable some ctags kinds

This patch is similar to the patch allowing us to enable/disable roles,
it just does this for ctags kinds. This can be useful when there is
a bug in a ctags parser causing incorrect parsing when certain kind is
enabled - which is what happens when parsing cython source files
with the 'z' flag enabled.

* Add LUA to the list of parsers returning full scope

* Enable/disable kinds in ctags based on whether we use them or not

Instead of enabling all kinds and then manually disabling them if
there are problems with them, we can enable/disable them based on whether
we actually use them - i.e., when they are mapped to something else
than tm_tag_undef_t.